requirements.txt
altgraph==0.17
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
bcrypt==3.2.0
beautifulsoup4==4.9.1
cachelib==0.1
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.3.1
cssselect==1.1.0
cycler==0.10.0
deprecated==1.2.10
distlib==0.3.1
django==3.1.1
emoji==0.6.0
et-xmlfile==1.0.1
fabric==2.5.0
filelock==3.0.12
future==0.18.2
howdoi==2.0.5
idna==2.10
invoke==1.4.1
isort==5.6.4
jdcal==1.4.1
keep==2.9
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.2
matplotlib==3.3.1
mccabe==0.6.1
mouseinfo==0.1.3
mysql-connector-python==8.0.21
numpy==1.19.1
openpyxl==3.0.5
pandas==1.1.1
paramiko==2.7.2
pefile==2019.4.18
pillow==7.2.0
pip==20.3.3
protobuf==3.13.0
pyautogui==0.9.50
pycparser==2.20
pyfirmata==1.1.0
pyftpdlib==1.5.6
pygetwindow==0.0.8
pygithub==1.53
pygments==2.6.1
pyinstaller-hooks-contrib==2020.9
pyinstaller==4.0
pyjwt==1.7.1
pylint==2.6.0
pymsgbox==1.0.8
pymysql==0.10.1
pynacl==1.4.0
pyparsing==2.4.7
pypdf2==1.26.0
pyperclip==1.8.0
pyquery==1.4.1
pyrect==0.1.4
pyscreeze==0.1.26
pyserial==3.4
python-dateutil==2.8.1
python-whois==0.7.3
pytweening==1.0.3
pytz==2020.1
pywin32-ctypes==0.2.0
requests==2.24.0
scipy==1.5.2
selenium==3.141.0
send2trash==1.5.0
setuptools==47.1.0
six==1.15.0
soupsieve==2.0.1
speedtest-cli==2.1.2
sqlparse==0.3.1
terminaltables==3.1.0
toml==0.10.1
twilio==6.45.4
urllib3==1.25.10
virtualenv==20.0.31
wikipedia==1.4.0
wrapt==1.12.1